+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJCYYPF6V5F12V3JQFPRZP2B"
+        "01KJD78967VV6NT9APZ3MGNF5W"
     )
     r"""The unique identifier of the data source"""