+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-26T12:31:53.721Z")
+    created: Optional[datetime] = parse_datetime("2026-02-26T14:56:55.966Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")