+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJCYYPFABK03PSF0C1CXEXM4"
+        "01KJD7896BXB7J00YKGYBS92Y1"
     )
     r"""The unique identifier of the data source"""